pyproject.toml
[tool.poetry]
name = "maple-lobbying"
version = "0.1.0"
description = "Scrape Massachusetts State House lobbying data"
authors = []
packages = [{ include = 'src' }]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
beautifulsoup4 = "^4.11.2"
certifi = "^2022.12.7"
charset-normalizer = "^3.0.1"
idna = "^3.4"
psycopg2 = "^2.9.5"
requests = "^2.28.2"
soupsieve = "^2.4"
urllib3 = "^1.26.14"
lxml = "^4.9.2"

[tool.poetry.group.synth.dependencies]
aws-cdk-lib = "^2.67.0"
constructs = "^10.1.268"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
